# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitnes Planer API",
        "description": "Dokumentacija koristeći PHP Atribute",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Lokalni API server"
        }
    ],
    "paths": {
...
        "/obroci": {
            "get": {
                "tags": [
                    "Obroci"
                ],
                "summary": "Lista obroka za datum",
                "operationId": "6a5ebd9d3003d6a0fbc65f7e12a81476",
                "parameters": [
                    {
                        "name": "datum",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Obroci"
                ],
                "summary": "Kreiranje novog obroka",
                "operationId": "a43c69614d22e4dad21ebbceabf36cde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "datum",
                                    "naziv",
                                    "namirnice"
                                ],
                                "properties": {
                                    "datum": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "naziv": {
                                        "type": "string"
                                    },
                                    "namirnice": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kreirano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/obroci/pregled": {
            "get": {
                "tags": [
                    "Obroci"
                ],
                "summary": "Grupisani pregled svih obroka po danima",
                "operationId": "9be99b46fcf7e32c8d9d1bfe608775e9",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/obrociPregled": {
            "get": {
                "tags": [
                    "Obroci"
                ],
                "summary": "Filtrirani pregled obroka po opsegu",
                "operationId": "eae7cd623ec3880b29403ee128583cc5",
                "parameters": [
                    {
                        "name": "od",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "do",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/obroci/{obrok}": {
            "put": {
                "tags": [
                    "Obroci"
                ],
                "summary": "Update obroka",
                "operationId": "29d4bdc629a6f9b29a000cb6f83d7252",
                "parameters": [
                    {
                        "name": "obrok",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ažurirano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Obroci"
                ],
                "summary": "Brisanje obroka",
                "operationId": "ea975d8bedba64cf2d09dbf5b24df02f",
                "parameters": [
                    {
                        "name": "obrok",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/obroci/{id}": {
            "get": {
                "tags": [
                    "Obroci"
                ],
                "summary": "Detalji obroka",
                "operationId": "5375a0dd0d96cd61ad00313122a7225c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}